+++ endstone/potion/__init__.py
@@ -3,6 +3,6 @@ import lazy_loader as lazy
 __getattr__, __dir__, __all__ = lazy.attach(
     "endstone._python",
     submod_attrs={
-        "potion": ["PotionType"],
+        "potion": ["Effect", "EffectType", "PotionType"],
     },
 )