+++ include/endstone/potion/effect_type.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <string>
+#include <string_view>
 
 #include "endstone/identifier.h"